Update and rename yaml.py to pubspec.py
